# backend/cmd/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API built with Golang using Clean Code architecture. \\nGithub: [https://github.com/shion0625/backend].",
    "title": "E-commerce Application Backend API",
    "contact": {
      "name": "For API Support",
      "email": "nikhilnarayanan623@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
  "paths": {
...
    "/admin/payment-method": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to change maximum price or block or unblock the payment method",
        "tags": ["Admin Payment"],
        "summary": "Update payment method (Admin)",
        "operationId": "UpdatePaymentMethod",
        "responses": {
          "200": {
            "description": "Successfully payment method updated",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "Invalid inputs",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to update payment method",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/payment-methods": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all payment methods",
        "tags": ["Admin Payment"],
        "summary": "Get payment methods (Admin)",
        "operationId": "GetAllPaymentMethodsAdmin",
        "responses": {
          "200": {
            "description": "Failed to retrieve payment methods",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Successfully retrieved all payment methods",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/carts/checkout/payment-select-page": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to render payment select page",
        "tags": ["User Payment"],
        "summary": "Render Payment Page (User)",
        "operationId": "CartOrderPaymentSelectPage",
        "responses": {
          "200": {
            "description": "Successfully rendered payment page",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to render payment page",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/carts/place-order/cod": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to place order for cash on delivery",
        "tags": ["User Payment"],
        "summary": "Place order  for COD (User)",
        "operationId": "PaymentCOD",
        "parameters": [
          {
            "type": "string",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully order placed for COD",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed place order for COD",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/carts/place-order/razorpay-checkout": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "API for user to create stripe payment",
        "tags": ["User Payment"],
        "summary": "Razorpay checkout (User)",
        "operationId": "RazorpayCheckout",
        "parameters": [
          {
            "type": "string",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully razorpay payment order created",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to make razorpay order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/carts/place-order/razorpay-verify": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for razorpay to callback backend for payment verification",
        "tags": ["User Payment"],
        "summary": "Razorpay verify (User)",
        "operationId": "RazorpayVerify",
        "parameters": [
          {
            "type": "string",
            "description": "Razorpay payment id",
            "name": "razorpay_order_id",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "Razorpay payment id",
            "name": "razorpay_payment_id",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "Razorpay signature",
            "name": "razorpay_signature",
            "in": "formData"
          },
          {
            "type": "string",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully razorpay payment verified",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "402": {
            "description": "Payment not approved",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to Approve order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/carts/place-order/stripe-checkout": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to create stripe payment",
        "tags": ["User Payment"],
        "summary": "Stripe checkout (User)",
        "operationId": "StripPaymentCheckout",
        "parameters": [
          {
            "type": "string",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully stripe payment order created",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to create stripe order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/carts/place-order/stripe-verify": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to callback backend after stripe payment for verification",
        "tags": ["User Payment"],
        "summary": "Stripe verify (User)",
        "operationId": "StripePaymentVeify",
        "parameters": [
          {
            "type": "string",
            "description": "Stripe payment ID",
            "name": "stripe_payment_id",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully stripe payment verified",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "402": {
            "description": "Payment not approved",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to Approve order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/payment-methods": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to get all payment methods",
        "tags": ["User Payment"],
        "summary": "Get payment methods (User)",
        "operationId": "GetAllPaymentMethodsUser",
        "responses": {
          "200": {
            "description": "Failed to retrieve payment methods",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Successfully retrieved all payment methods",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "error": {},
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}